Improve gaussian fit by tunning initial std dev
Improve gaussian fit by tunning initial standard deviation.
Gaussian fits where not correctly done. The initial parameters
need an initial tunning. A default value of 1.0 was used for the
standard deviation. This was not allowing to find good fits for
Gaussian data. We have tunned this value at the beginning by
taking into account the range of the abscissa data.

Now the fits are much more correct, and better represents the
real data.